refactor: update trigger characters and keyword patterns
Don't trigger the completion by any keywords (use a pattern that is not
likely to be triggered.). only trigger on the given characters. This is
because candidates returned by LLMs are easily filtered out by cmp due
to that LLM oftern returns candidates contains the full content in
current line before the cursor.