feat: cancel completion jobs to run new ones
This patch assumes that whenever complete() is triggered, currently
running completion requests are obsolete and their result should not be
processed anymore.

Therefore, this patch adds a mechanism to keep track of running
completion jobs and terminates them, if they are still running when
complete() is called again.

This fixes a race condition where old complete requests might still
contribute to the suggestions even though they are not up to date
anymore.

This reduces LLM load even when throttle and debounce are set to 0
because there are always only n_completions jobs running at the same
time.